requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-core
chromadb
sentence-transformers
pymupdf
mistralai
python-dotenv
